requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
spacy>=3.7.0
python-multipart>=0.0.9
aiofiles>=23.2.1
websockets>=12.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.1.0
requests>=2.31.0
